refactor: split ZUpdateAvatarAssistantInputSchema
